pyproject.toml
[project]
name = "simple-code-reviewer"
version = "0.1.0"
description = "Simple AI chat application using Gradio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-secret-manager>=2.22.0",
    "google-generativeai>=0.8.3",
    "gradio>=3.50.2",
]
